+++ src/SimpleW.Templates/templates/SimpleW.AspNetCore/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using Microsoft.Extensions.Hosting;
+using SimpleW;
 using SimpleW.Helper.Hosting;
 using SimpleW.Helper.Razor;
 
@@ -37,7 +38,10 @@ namespace ModuleName {
                     });
 
                     // subscribe to events
-                    server.OnStarted(s => {
+                    server.OnStateChanged((s, state) => {
+                        if (state != SimpleWServerState.Started) {
+                            return;
+                        }
                         Process.Start(new ProcessStartInfo {
                             FileName = $"http://localhost:{s.Port}/home/index?name=Chris",
                             UseShellExecute = true